Define explicit primary keys and regenerate DBIx classes
In order to retrieve the last_insert_id, the Schema needs to know what
field is the primary key: I had to add this information to the schema in
sql/tables.sql for every table.

I added the retrieval of these last_insert_id in all the methods that
need to return such ids.