Add species_id in sample table
Instead of relying on a vague link between the taxon_id+strain between
the sample and the species table, it is better to define a parameter for
each sample with a clear associated species (prod_name).

The reason is that some samples have no strain, or a strangely named
strain (e.g. SINENSIS_R1 for SINENSIS). In those cases, it is possible
to guess the best match, but it should always be possible to override
this decision.

If a sample doesn't correctly match, there are 2 choices:
- add a species entry
- change the sample.specie_id to match the correct species