Add species guessing for samples
Samples now have a species_id, but matching the taxon_id+strain to the
species table is not straightforward: some samples don't have a strain,
and some of them use a differently named strain.

The script now guesses the species_id. In some cases there are several
candidates possible for a species (strains with the same taxon_id but
different strains and production_name, e.g. An. stephensiI). In that
case, the script chooses one the species and warns the user that the
species chosen may not be the right one.

The species table now allow several entries with the same
production_name to have aliases, for example to define type strains
(leaving the strain column empty will ensure that the script picks it by
default).