Add tests for species insertion
For those tests, a temporary database is created, and deleted at the end
of the tests. A configuration file is therefore necessary (db_conf.json,
empty by default). The name of the database has a random number appended
to it in order to avoid a collision with an existing database (the
script will die if it could not create it).