Add function to get a private study (without SRA)
Some studies may not have an SRA (only some fastq files), so the
add_private_study function can import RNA-seq metadata for those
studies. They need to be described in a hierarchical format like SRA,
see the json example in the perldoc.

The easiest way to import is with a json file.

NB: those data don't have an SRA, so they use the field run_private_acc
instead of run_sra_acc (for the run table).