Add a track for each sample, and a function to get all new tracks
A new track to be created is simply a row of the table track that does
not have a file_id associated with it.

In order to retrieve the tracks data, a join needs to be done between
the tables track - sample - strain - species. For that I had to add
relationships between those tables.

New tests for the tracks added.