Add table sra_track
Instead of linking tracks directly to samples, it is better to have a
linking table, which would allow linking to any SRA table. This would
also allow defining a track with several SRA ids, which means they would
be merged.

The merge level needs to be defined in the tracks table though, in order
to give clear instructions to the RNAseq pipeline.