Add linker between drupal_node and tracks
Instead of linking to experiments or studies, drupal_nodes are better
linked to tracks. More specifically, we should have a drupal node for
every active track. The linker makes it possible to group tracks in an
upper level drupal_node.

Updated the add_track functions to automatically create a linked
drupal_node.