Update DBIx schema and tinker tracks files recognition for new schema
The new schema means that the tracks lost their file_id which was used
to detemine wether they had to be aligned.

So I had to add a preparation phase to retrieve the files, check which
ones exist, and determine if the track can be aligned or not.

This change also allows to get the fastq files associated with a private
study for them to be used by the RNAseq pipeline.