Export both merge_level and merge_id for the pipeline
In the cases where the merging is special, i.e. not 'sample' or 'study',
the merge_level 'taxon' is used.

Ideally the merge_level and merge_id should be precomputed and stored in
the database. I updated the schema for that but for now the
merge_id/level are computed on the fly when exporting the list of tracks
to create for the pipeline.