Remove paths from imported pipeline commands
The files themselves will be moved anyway to a different directory
structure, so it makes little sense to import their own path. And
removing the paths still allows to reproduce the processes.

So the actual path of the files will have to be reconstructed based on a
root dir that the user will have to provide.