Add script to merge and annotate tracks and bundles
In replacement of merge_tracks, and drupal_tracks. Both tracks and
bundles can now be merged and annotated with data a json file.

The script first finds the tracks/bundles from a list of SRAs from the
json file. Then the corresponding tracks can be merged (if asked to), as
well as annotated.

A third option for bundles allows to copy the text and title of a track,
if the bundle is only made of 1 track.