Ignore pipeline taxonomy restriction: we already control that
Some SRA ids can have a taxonomic id diffrent from what is in the
RNAseqDB. This is the case for example if the taxon is a more generic
one (e.g. Glossina morsitans morsitans = Glossina morsitans).
We still want this SRA experiment to be aligned against the genome
(that's why we put it in the DB), so we have to bypass the RNAseq
pipeline taxonomic restriction (it has a dedicated parameter for that).