Update tracks/bundles can use a species name in case of ambiguity
The json file contains a list of SRA ids, but sometimes a study contains
several species, and we separate the tracks for each species. In order
to not mix the species when merging, and to not have to write down all
the individual runs/experiments/samples, it is easier to give the study
id and the species name.